pdfium_fast/integration_tests/requirements.txt
# PDFium Test Suite Requirements

# Core testing
pytest>=7.0.0
pytest-html>=3.1.1
pytest-timeout>=2.1.0

# Telemetry & statistics
pandas>=1.5.0
psutil>=5.9.0
filelock>=3.12.0

# Enhanced validation (optional)
openai>=1.0.0  # LLM analysis
Pillow>=9.0.0  # Pixel comparison
numpy>=1.20.0  # Image arrays
